fix --check-only panic
