+++ internal/pool/types.go
@@ -1,6 +1,9 @@
 package pool
 
-import "github.com/chainreactors/parsers"
+import (
+	"github.com/chainreactors/parsers"
+	"github.com/chainreactors/spray/pkg"
+)
 
 func newUnit(path string, source parsers.SpraySource) *Unit {
 	return &Unit{path: path, source: source}
@@ -17,4 +20,17 @@ type Unit struct {
 	retry    int
 	frontUrl string
 	depth    int // redirect depth
+}
+
+func NewBaselines() *Baselines {
+	return &Baselines{
+		baselines: map[int]*pkg.Baseline{},
+	}
+}
+
+type Baselines struct {
+	FailedBaselines []*pkg.Baseline
+	random          *pkg.Baseline
+	index           *pkg.Baseline
+	baselines       map[int]*pkg.Baseline
 }
